Adding more profile info to results and script

+++ scripts/selectivity_exp.py
@@ -6,15 +6,15 @@ binary = "build/release/main_cu"
 
 default_filter_size = 536870912    #64  MiB
 default_streams = 4
-default_probe_size = 419430400     #400 M keys
-default_build_size = 4194304   	   #4   M keys
+default_probe_size = 1000#419430400     #400 M keys
+default_build_size = 1000#4194304   	   #4   M keys
 default_num_threads = 32
 default_gpu_morsel_size = 16777216 #16  M keys
 default_cpu_morsel_size = 16384	   #16  K keys
 default_gpu_devices = 0
 default_selectivity = 1
 
-selectivities = [1, 10, 20, 30, 40, 50, 60, 70, 80, 90]
+selectivities = [1 for y in range(10)]
 
 def syscall(cmd):
 	print(cmd)
@@ -36,6 +36,7 @@ def run_test(fname = None, probe_size = None, streams = None, filter_size = None
 	if os.path.isfile(os.path.join('results', fname)):
 		return
 
+	# Execute Experiment
 	syscall("""${BINARY} --filter_size=${FILTER_SIZE} --probe_size=${PROBE_SIZE} --build_size=${BUILD_SIZE} --gpu_morsel_size=${GPU_MORSEL_SIZE} --cpu_morsel_size=${CPU_MORSEL_SIZE} --gpu=${DEVICES} --selectivity=${SELECTIVITY} --num_threads=${THREADS}""".replace(
 		"${BINARY}", binary).replace(
 		"${FILTER_SIZE}", str(filter_size)).replace(
@@ -46,18 +47,22 @@ def run_test(fname = None, probe_size = None, streams = None, filter_size = None
 		"${DEVICES}", str(gpu_devices)).replace(
 		"${SELECTIVITY}", str(selectivity)).replace(
 		"${THREADS}", str(threads)))
-	os.system('mv results.csv %s' % os.path.join('results', fname))
+	# We include the header in the first time
+	if not os.path.isfile(os.path.join('results', fname)):
+		os.system('mv results.csv %s' % os.path.join('results', fname))
+	else:
+		os.system('sed -n \'2,3p\' results.csv >> %s' % os.path.join('results', fname))
 
 
 os.system('make')
 os.system('mkdir -p results')
 
-#os.system('mkdir -p results/options')
-#for opt in options:
-#	run_test(fname="options/results-%s.csv" % (opt.replace(" ", "")), options = opt)
-
 
 os.system('mkdir -p results/selectivity')
 for selectivity in selectivities:
-	run_test(fname="selectivities/results-selectivity%s.csv" % (str(selectivity)), selectivity=selectivity)
+	run_test(fname="selectivity/results-selectivity_cpu.csv" % (str(selectivity)), selectivity=selectivity)
+
+os.system('mkdir -p results/selectivity')
+for selectivity in selectivities:
+	run_test(fname="selectivity/results-selectivity_gpu.csv" % (str(selectivity)), selectivity=selectivity, gpu_devices=1)
